Add `replace_config` property.
This allows us to be more flexible about how to set the default configuration
file in the `module_name` file. By default the file won't be changed, but if
you set this property to a string or regular expression, any text in the file
that matches will be replaced with the name of the context configuration file.

That takes care of the most important changes. Next I'll have to add some
tests.